+++ src/components/scenes/SceneList.jsx
@@ -1,7 +1,8 @@
-import { useRef } from 'react';
+import { Fragment, useRef } from 'react';
 
 import { useSceneIds } from '@/stores/document';
 import { useDraggableContainerAutoScroll } from '@/stores/draggable';
+import { useCurrentCursor } from '@/stores/user';
 
 import DocumentTitle from '../shotlist/DocumentTitle';
 import NewSceneEntry from './NewSceneEntry';
@@ -15,14 +16,28 @@ export default function SceneList({ documentId }) {
   const containerRef = useRef(null);
   const sceneIds = useSceneIds(documentId);
   useDraggableContainerAutoScroll(containerRef);
+
+  const userCursor = useCurrentCursor();
+  const activeSceneId = userCursor.sceneId;
+  const hasActiveScene = Boolean(activeSceneId);
+
   return (
     <article
       className="w-full h-full overflow-x-hidden overflow-y-auto py-20"
       ref={containerRef}>
       <DocumentTitle documentId={documentId} />
-      {sceneIds.map((sceneId) => (
-        <SceneEntry key={sceneId} documentId={documentId} sceneId={sceneId} />
-      ))}
+      {sceneIds.map(
+        (sceneId) =>
+          (!hasActiveScene || sceneId === activeSceneId) && (
+            <Fragment key={`scene-${sceneId}`}>
+              <SceneEntry
+                key={sceneId}
+                documentId={documentId}
+                sceneId={sceneId}
+              />
+            </Fragment>
+          ),
+      )}
       <NewSceneEntry documentId={documentId} />
     </article>
   );